+++ EVCar/src/main/java/com/company/service/ReplyService.java
@@ -10,14 +10,14 @@ public interface ReplyService {
 
 	public int register(ReplyVO vo);
 
-	public ReplyVO get(Long rno);
+	public ReplyVO get(int rno);
 
 	public int modify(ReplyVO vo);
 
-	public int remove(Long rno);
+	public int remove(int rno);
 
-	public List<ReplyVO> getList(Criteria cri, Long bno);
+	public List<ReplyVO> getList(Criteria cri, int bno);
 	
-	public ReplyPageDTO getListPage(Criteria cri, Long bno);
+	public ReplyPageDTO getListPage(Criteria cri, int bno);
 
 }
